+++ src/NukiLock.h
@@ -169,7 +169,6 @@ class NukiLock : public Nuki::NukiBle {
      */
     Nuki::CmdResult enableAutoUpdate(const bool enable);
 
-
     /**
      * @brief Sets the lock ability to pair with other devices (can be used to prevent unauthorized pairing)
      * Gets the current config from the lock, updates the pairing parameter and sends the new config to the lock via BLE
@@ -180,6 +179,10 @@ class NukiLock : public Nuki::NukiBle {
      */
     Nuki::CmdResult enablePairing(const bool enable);
 
+    /**
+     * @brief Gets the lock current config wrt pairing with other devices
+     */
+    bool pairingEnabled();
 
     /**
      * @brief Gets the current config from the lock, updates the whether or not the flashing
@@ -272,6 +275,14 @@ class NukiLock : public Nuki::NukiBle {
                                        const bool totalCount);
 
     /**
+     * @brief Request the lock via BLE to send the authorization entries
+     *
+     * @param offset Startindex of first log msg to be send
+     * @param count The number of log entries to be read, starting at the specified start index.
+     */
+    Nuki::CmdResult retrieveAuthorizationEntries(const uint16_t offset, const uint16_t count);
+
+    /**
      * @brief Returns battery critical state parsed from the battery state byte (battery critical byte)
      *
      * Note that `retrieveOpenerState()` needs to be called first to retrieve the needed data
@@ -321,6 +332,7 @@ class NukiLock : public Nuki::NukiBle {
     BatteryReport batteryReport;
     std::list<TimeControlEntry> listOfTimeControlEntries;
     std::list<LogEntry> listOfLogEntries;
+    std::list<AuthorizationEntry> listOfAuthorizationEntries;
 
     Config config;
     AdvancedConfig advancedConfig;