+++ src/NukiLockUtils.cpp
@@ -534,8 +534,8 @@ void logKeyturnerState(KeyTurnerState keyTurnerState) {
   logCompletionStatus(keyTurnerState.lastLockActionCompletionStatus);
   log_d("doorSensorState: %d", keyTurnerState.doorSensorState);
   log_d("nightModeActive: %d", keyTurnerState.nightModeActive);
-  log_d("Keypad bat critical feature supported: %d", keyTurnerState.AccessoryBatteryState & 1);
-  log_d("Keypad Battery Critical: %d", keyTurnerState.AccessoryBatteryState & 2);
+  log_d("Keypad bat critical feature supported: %d", keyTurnerState.accessoryBatteryState & 1);
+  log_d("Keypad Battery Critical: %d", keyTurnerState.accessoryBatteryState & 2);
   #endif
 }